feat(freestream): refactor file upload, retrieval configuration, and custom callback handlers
Continuing the cleanup of `main.py`, I moved the custom callbacks to the utility module and extensively enhanced their docstrings for better clarity
